reproduce accuracy of fastJLT on mediamill

+++ src/MatrixLoader/MediaMillMatrixLoader.cpp
@@ -15,11 +15,11 @@ void AMMBench::MediaMillMatrixLoader::generateAB() {
 
   torch::jit::script::Module tensors = torch::jit::load(filePath);
   // A, B already normalized
-  A = tensors.attr("A").toTensor(); // 120*43907
-  B = tensors.attr("B").toTensor(); // 101*43907
+  A = tensors.attr("A").toTensor().to(torch::kFloat); // 120*43907
+  B = tensors.attr("B").toTensor().to(torch::kFloat); // 101*43907
 
-  At = A.t().contiguous(); // 43907*120
-  Bt = B.t().contiguous(); // 43907*101
+  At = A.t().contiguous().to(torch::kFloat); // 43907*120
+  Bt = B.t().contiguous().to(torch::kFloat); // 43907*101
 
   std::cout << "Maximum Value: " << A.max().item<float>() << std::endl;
   std::cout << "Mean Value: " << A.mean().item<float>() << std::endl;
